# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Trip Planner API",
    "description": "Collaborative trip planning, itineraries, routes, expenses, and settlements.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "actorUserId": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/routes/trip/{tripId}": {
      "get": {
        "summary": "List trip routes",
        "tags": [
          "routes"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "query",
            "name": "actorUserId",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "tripId",
            "required": true
          }
        ],
        "security": [
          {
            "actorUserId": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "tripId": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "mode": {
                        "type": "string",
                        "enum": [
                          "DRIVE",
                          "WALK",
                          "BIKE",
                          "TRANSIT"
                        ]
                      },
                      "startsAt": {
                        "nullable": true,
                        "anyOf": [
                          {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                          },
                          {
                            "type": "string",
                            "format": "date-time"
                          }
                        ]
                      },
                      "endsAt": {
                        "nullable": true,
                        "anyOf": [
                          {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                          },
                          {
                            "type": "string",
                            "format": "date-time"
                          }
                        ]
                      },
                      "locked": {
                        "type": "boolean"
                      },
                      "createdAt": {
                        "anyOf": [
                          {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                          },
                          {
                            "type": "string",
                            "format": "date-time"
                          }
                        ]
                      },
                      "legs": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "routePlanId": {
                              "type": "string"
                            },
                            "fromPlaceId": {
                              "type": "string"
                            },
                            "toPlaceId": {
                              "type": "string"
                            },
                            "order": {
                              "type": "integer",
                              "minimum": -9007199254740991,
                              "maximum": 9007199254740991
                            },
                            "distanceMeters": {
                              "nullable": true,
                              "type": "integer",
                              "minimum": -9007199254740991,
                              "maximum": 9007199254740991
                            },
                            "durationSeconds": {
                              "nullable": true,
                              "type": "integer",
                              "minimum": -9007199254740991,
                              "maximum": 9007199254740991
                            },
                            "encodedPolyline": {
                              "nullable": true,
                              "type": "string"
                            },
                            "provider": {
                              "nullable": true,
                              "type": "string"
                            },
                            "fromPlace": {},
                            "toPlace": {}
                          },
                          "required": [
                            "id",
                            "routePlanId",
                            "fromPlaceId",
                            "toPlaceId",
                            "order",
                            "distanceMeters",
                            "durationSeconds",
                            "encodedPolyline",
                            "provider"
                          ],
                          "additionalProperties": false
                        }
                      }
                    },
                    "required": [
                      "id",
                      "tripId",
                      "name",
                      "mode",
                      "startsAt",
                      "endsAt",
                      "locked",
                      "createdAt",
                      "legs"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
    "/routes": {
      "post": {
        "summary": "Create route plan from ordered places",
        "tags": [
          "routes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "actorUserId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "tripId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "mode": {
                    "default": "DRIVE",
                    "type": "string",
                    "enum": [
                      "DRIVE",
                      "WALK",
                      "BIKE",
                      "TRANSIT"
                    ]
                  },
                  "startsAt": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                  },
                  "endsAt": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                  },
                  "placeIds": {
                    "minItems": 2,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1
                    }
                  }
                },
                "required": [
                  "tripId",
                  "name",
                  "placeIds"
                ]
              }
            }
          }
        },
        "security": [
          {
            "actorUserId": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "tripId": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "DRIVE",
                        "WALK",
                        "BIKE",
                        "TRANSIT"
                      ]
                    },
                    "startsAt": {
                      "nullable": true,
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        },
                        {
                          "type": "string",
                          "format": "date-time"
                        }
                      ]
                    },
                    "endsAt": {
                      "nullable": true,
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        },
                        {
                          "type": "string",
                          "format": "date-time"
                        }
                      ]
                    },
                    "locked": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        },
                        {
                          "type": "string",
                          "format": "date-time"
                        }
                      ]
                    },
                    "legs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "routePlanId": {
                            "type": "string"
                          },
                          "fromPlaceId": {
                            "type": "string"
                          },
                          "toPlaceId": {
                            "type": "string"
                          },
                          "order": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "distanceMeters": {
                            "nullable": true,
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "durationSeconds": {
                            "nullable": true,
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "encodedPolyline": {
                            "nullable": true,
                            "type": "string"
                          },
                          "provider": {
                            "nullable": true,
                            "type": "string"
                          },
                          "fromPlace": {},
                          "toPlace": {}
                        },
                        "required": [
                          "id",
                          "routePlanId",
                          "fromPlaceId",
                          "toPlaceId",
                          "order",
                          "distanceMeters",
                          "durationSeconds",
                          "encodedPolyline",
                          "provider"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "id",
                    "tripId",
                    "name",
                    "mode",
                    "startsAt",
                    "endsAt",
                    "locked",
                    "createdAt",
                    "legs"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/expenses/trip/{tripId}": {
      "get": {
        "summary": "List trip expenses",
        "tags": [
          "expenses"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "query",
            "name": "actorUserId",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "tripId",
            "required": true
          }
        ],
        "security": [
          {
            "actorUserId": []
          }
        ],
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "description": "JSON response"
                }
              }
            }
          }
        }
      }
    },
    "/expenses": {
      "post": {
        "summary": "Create expense",
        "tags": [
          "expenses"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "actorUserId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "tripId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "paidById": {
                    "type": "string",
                    "minLength": 1
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1
                  },
                  "amount": {
                    "type": "number",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "currency": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 3
                  },
                  "originalAmount": {
                    "type": "number",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "originalCurrency": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 3
                  },
                  "exchangeRate": {
                    "type": "number",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "exchangeDate": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                  },
                  "itineraryStopId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "splitType": {
                    "default": "EQUAL",
                    "type": "string",
                    "enum": [
                      "EQUAL",
                      "CUSTOM"
                    ]
                  },
                  "splitAllTripMembers": {
                    "default": false,
                    "type": "boolean"
                  },
                  "splitUserIds": {
                    "minItems": 1,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "customSplits": {
                    "minItems": 1,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "userId": {
                          "type": "string",
                          "minLength": 1
                        },
                        "amount": {
                          "type": "number",
                          "minimum": 0,
                          "exclusiveMinimum": true
                        }
                      },
                      "required": [
                        "userId",
                        "amount"
                      ]
                    }
                  }
                },
                "required": [
                  "tripId",
                  "paidById",
                  "title",
                  "amount"
                ]
              }
            }
          }
        },
        "security": [
          {
            "actorUserId": []
          }
        ],
        "responses": {
          "201": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "description": "JSON response"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ]
}